Refactor Jupyter magic unit tests (#3599)
* move fixture to conftest


* move fixture


* move dummy function out


* refactor multiline functions


* remove unncessary mock


* remove functions


* update docstring


---------
